pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jira-mcp-server"
version = "1.0.0"
description = "MCP server for Jira & Confluence — 53 tools for issue management, sprint planning, board management, and Confluence page management"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
  { name = "Muhammad Abdul Halim", email = "abdul.halim@brainstation-23.com" }
]
keywords = ["mcp", "jira", "confluence", "atlassian", "model-context-protocol", "ai", "claude"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries",
  "Topic :: Internet :: WWW/HTTP",
]
requires-python = ">=3.10"
dependencies = [
  "mcp>=1.0.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/halim-23/jira-mcp-server"
Repository = "https://github.com/halim-23/jira-mcp-server"
Issues = "https://github.com/halim-23/jira-mcp-server/issues"
Changelog = "https://github.com/halim-23/jira-mcp-server/blob/main/CHANGELOG.md"

[project.scripts]
jira-mcp-server = "jira_mcp_server.__main__:run"

[tool.hatch.build.targets.wheel]
packages = ["jira_mcp_server"]

[tool.hatch.build.targets.sdist]
include = [
  "jira_mcp_server/",
  "README.md",
  "LICENSE",
  "CHANGELOG.md",
  "pyproject.toml",
]
